requirements.txt
websocket-client==0.57.0
python-dotenv==0.15.0
